Set list title endpoint and database

+++ migrations/20231117224650_update_list_titles_table.js
@@ -3,18 +3,22 @@
  * @returns { Promise<void> }
  */
 exports.up = function (knex) {
-  return knex.schema.createTable("list_titles", (table) => {
-    table.increments("id").primary();
-    table.timestamp("date").defaultTo(knex.fn.now());
-    table.string("title").notNullable();
-    table
-      .integer("notebook_id")
-      .unsigned()
-      .references("id") // Explicitly reference the column without the table name prefix
-      .inTable("notebooks") // Explicitly specify the table name without relying on default resolution
-      .onUpdate("CASCADE")
-      .onDelete("CASCADE")
-      .defaultTo(1);
+  return knex.schema.hasTable("list_titles").then(function (exists) {
+    if (!exists) {
+      return knex.schema.createTable("list_titles", (table) => {
+        table.increments("id").primary();
+        table.timestamp("date").defaultTo(knex.fn.now());
+        table.string("title").notNullable();
+        table
+          .integer("notebook_id")
+          .unsigned()
+          .references("id") // Explicitly reference the column without the table name prefix
+          .inTable("notebooks") // Explicitly specify the table name without relying on default resolution
+          .onUpdate("CASCADE")
+          .onDelete("CASCADE")
+          .defaultTo(1);
+      });
+    }
   });
 };
